Create azide_rename.py
The program is designed to have a few fully parameterized novel amino acids, using the CuAAC alkyne and azide to emulate common experimental modifications. Usually the novel amino acid has been fully parameterized with a terminal alkyne. The incoming protein will be patched in at an azide to the alkyne. To patch the protein, the azide nitrogens need to have consistent naming. This program renames the azide nitrogens to N11, N12, and N13.