Update Internal_coordinate_writer.cpp
This program writes the internal coordinates to a file, used in conjunction with the .str file from the CGENFF website, automates the topology writing process for the CHARMM force field. I run my programs on a MAC, but CGENFF allows use of their program for free to universities, however their code only works on Linux machines. I am writing an alternative program for this purpose on a different computer that will include the CGENFF program, essentially automating the entire process