Update rotate_protein.py
This takes the output from the create_rotation_matrix program and implements the matrix to rotate the protein to the correct position. The rotation_matrix.txt file is created by the create_rotation_matrix.cpp. The pdb input file (P2.pdb) is created by the orient_protein.py program. The output file (P2_nu.pdb) will be used in the next step.